Restructure Router (1)
Before:
App>Router>HeaderUser>{Header, Drawer, UserPage>Box>Switch}

Now
App>Router>Switch>HeaderUser

And the desired page is passed as a prop of HeaderUser
